+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 --------------------
 - Fix #315: Fix participants list settings visibility
 - Fix #316: Fix invitation of space members
+- Enh #261: Filter by event types
 
 1.2.3 (May 30, 2022)
 --------------------